requirements.txt
google-genai
cryptography
beautifulsoup4
requests